% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalSpectra.R
\name{getFunctionalSpectrum}
\alias{getFunctionalSpectrum}
\title{Generate Functional Spectrum}
\usage{
getFunctionalSpectrum(
  expressionProfile,
  geneSets = "MSigDBv7",
  show_progress = F,
  No = NULL
)
}
\arguments{
\item{expressionProfile}{a named numeric vector containing gene expression profile}

\item{geneSets}{a List containing gene sets (default: MSiDB v6)}
}
\value{
a numeric vector containing functional spectrum
}
\description{
This function generates functional spectrum for a single gene expression profile.
}
\note{
You can generate the reference expression profile from your previous data or public data, which is the same(similiar) cancer type and platform.
In DeepCC we also prepared average expression profiles of each cancer types in TCGA project as references. To use them, just use the TCGA identifier (COADREAD, BRCA, OV, etc.) to indicate the cancer type.
If your single sample is microarray data, we strongly sugguest turn the parameter \code{inverseRescale} on, since TCGA is RNA-Seq, which has very small expression value for low expressed genes, compared with microarray.
}
\examples{
getFunctionalSpectrum(ep)
}
\seealso{
\code{\link{get_functional_spectra}} for a batch of gene expression profiles.
}
