% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalSpectra.R
\name{multi_sample_ssgsea}
\alias{multi_sample_ssgsea}
\title{Generate Functional Spectrum for multi samples}
\usage{
multi_sample_ssgsea(
  eps,
  geneSet = "MSigDBv7",
  cores = parallel::detectCores() - 2,
  show_progress = FALSE
)
}
\arguments{
\item{eps}{a data.frame containing gene expression profiles (each row presents one sample)}

\item{cores}{a integer indicating cpu cores used in parallel computing (default = all cores -2 )}

\item{show_progress}{a bool value, to show the running progress, default: FALSE}

\item{geneSets}{a List containing gene sets (default: MSiDB v6)}
}
\value{
a numeric vector containing functional spectrum
}
\description{
This function generates functional spectrum for a single gene expression profile for multiple samples.
}
\note{
You can generate the reference expression profile from your previous data or public data, which is the same(similiar) cancer type and platform.
In DeepCC we also prepared average expression profiles of each cancer types in TCGA project as references. To use them, just use the TCGA identifier (COADREAD, BRCA, OV, etc.) to indicate the cancer type.
If your single sample is microarray data, we strongly sugguest turn the parameter \code{inverseRescale} on, since TCGA is RNA-Seq, which has very small expression value for low expressed genes, compared with microarray.
}
\examples{
multi_sample_ssgsea(ep, geneSet = "MsigDBv7", show_progress = T)
}
\seealso{
\code{\link{get_functional_spectra}} for a batch of gene expression profiles.
}
