% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalSpectra.R
\name{single_newCalcEnrichmentScore}
\alias{single_newCalcEnrichmentScore}
\title{Calculate single sample Enrichment Score}
\usage{
single_newCalcEnrichmentScore(geneList, geneSet)
}
\arguments{
\item{geneList}{a named vector containing the values of gene expression}

\item{geneSet}{a vector containing genes to represent a gene set}
}
\value{
a numeric indicating enrichment score
}
\description{
This function calculates enrichment score of a gene list on a specific gene set.
}
\examples{
single_newCalcEnrichmentScore(geneList, geneSet)
}
